Bump `abscissa`, `tokio`, `tendermint-rpc`, `warp`
Upgrades the following crates:

- `abscissa*` to v0.6.0-pre.1
- `tokio` to v1
- `tendermint-rpc` to v0.19
- `warp` to v0.3

This is a pretty major refactor which draws some inspiration from the
structure of Delphi.

Previously Sagan was keeping the collector state in the `APP` state
itself, however, `abscissa*` v0.6.0-pre.1 makes this state immutable.

Instead of that, the collector state is now kept in a `collector::State`
struct which performs largely the same function. This seems like a good
refactoring per the single responsibility principle (SRP).

This commit also extracts a request router which is much closer to the
one currently used in Delphi into `collector::Router`.