Mintscan integration
Initial support for using Mintscan for black box monitoring, providing a
source for chain state and validator uptime that isn't our own
infrastructure.

Adds a "poller" subsystem to the collector for data sources that are
polled instead of pushed.

Refactors configuration to support additional information about
Tendermint networks, including a Mintscan API endpoint and the
validator's operator address.

Adds an initial Mintscan poller which checks the chain status endpoint
and, if the validator address is configured, the validator uptime
endpoint for the configured address.

If the validator's last signed height exceeds a certain threshold away
from the chain tip, it presently emits an `error!` log event.

The longer term goal is to use the `datadog` crate to page us when this
happens, but that work is deliberately left to a followup PR.