Refactor using Tower (#181)
Tower is a library of modular and reusable components for building
robust networking clients and servers:

https://docs.rs/tower/

The main thing Tower provides in the context of Sagan is a channel-based
abstraction which allows several other client tasks to talk to a common
service. We need this because events can come in via the Warp-based
HTTP API or via polling of external services (i.e. `collector::Poller`)

This commit does a hack-and-slash replacement of `collector::State`
with a new `Collector` implemented as a `tower::Service`.

Right now it's only used to handle requests coming in via Warp, but the
next step is to make the collector accessible to `collector::Poller` as
well and allow it to send events to the service.