Report events from `Poller` to `Collector`
Has the `mintscan::Poller` report data it collects to the `Collector`
service in the form of `Request::PollInfo` messages.

This code is a bit handwavy as we're feeling out the requirements as we
go, and doesn't yet persist the reported events in any way, but it does
log them and provide a path towards keeping the business logic around
things like alerting in the `Collector` rather than the individual
pollers.